requirements.txt
yt-dlp
rich
pyfiglet
pyperclip
requests
pillow